Backend/requirements.txt
fastapi==0.70.0
python-dotenv==0.19.2
uvicorn==0.15.0
python-multipart==0.0.5
nest-asyncio==1.5.1
aiohttp==3.7.4.post0